# openapi.yaml
openapi: 3.1.0
info:
  title: 27001certified API
  version: 1.0.0
  description: |
    Backend API for 27001certified.app — ISO 27001:2022 AI-Certification Portal.
    All protected endpoints require a JWT Bearer token unless marked as public.
  contact:
    name: 27001certified Platform
    url: https://27001certified.app
servers:
  - url: https://api.27001certified.app
    description: Production
  - url: https://api-staging.27001certified.app
    description: Staging
  - url: http://localhost:5000
    description: Local development

# ────────────────────────────────────────────────────────────
#  SECURITY SCHEMES
# ────────────────────────────────────────────────────────────
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # ── SHARED SCHEMAS ─────────────────────────────────────────
  schemas:
...
    User:
      type: object
      properties:
        id:            { type: string, format: uuid }
        full_name:     { type: string }
        email:         { type: string, format: email }
        company:       { type: string, nullable: true }
        job_role:      { type: string, nullable: true }
        country:       { type: string, nullable: true }
        email_verified:{ type: boolean }
        role:          { type: string, enum: [user, admin] }
        created_at:    { type: string, format: date-time }
...
    ExamSession:
      type: object
      properties:
        id:            { type: string, format: uuid }
        user_id:       { type: string, format: uuid }
        track:         { type: string, enum: [associate, professional] }
        status:
          type: string
          enum: [generating, ready, active, completed, expired, aborted]
        score:         { type: integer, nullable: true }
        passed:        { type: boolean, nullable: true }
        tab_violations:{ type: integer }
        started_at:    { type: string, format: date-time, nullable: true }
        completed_at:  { type: string, format: date-time, nullable: true }
        expires_at:    { type: string, format: date-time, nullable: true }
        created_at:    { type: string, format: date-time }
    Certificate:
      type: object
      properties:
        id:               { type: string, format: uuid }
        certificate_id:   { type: string, example: "ISO27-2026-A1B2C3D4" }
        user_id:          { type: string, format: uuid }
        exam_session_id:  { type: string, format: uuid }
        track:            { type: string, enum: [associate, professional] }
        full_name:        { type: string }
        score:            { type: integer }
        awarded_on:       { type: string, format: date-time }
        r2_url_landscape: { type: string, format: uri, nullable: true }
        r2_url_square:    { type: string, format: uri, nullable: true }
        revoked:          { type: boolean }
        revoked_reason:   { type: string, nullable: true }
...
    Error:
      type: object
      required: [error]
      properties:
        error:   { type: string }
        message: { type: string }
        code:    { type: string }

    # ── Pagination ─────────────────────────────────────────
    Pagination:
      type: object
      properties:
        total: { type: integer }
        page:  { type: integer }
        pages: { type: integer }
        limit: { type: integer }

  # ── RESPONSES ──────────────────────────────────────────────
  responses:
    Unauthorized:
      description: Missing or invalid JWT
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Forbidden:
      description: Insufficient role permissions
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
security:
  - BearerAuth: []

# ────────────────────────────────────────────────────────────
#  PATHS
# ────────────────────────────────────────────────────────────
paths:
...
  /api/admin/users:
    get:
      summary: List users with search/filter/pagination
      tags: [Admin]
      parameters:
        - in: query
          name: page
          schema: { type: integer, default: 1 }
        - in: query
          name: limit
          schema: { type: integer, default: 25 }
        - in: query
          name: search
          schema: { type: string }
        - in: query
          name: status
          schema: { type: string, enum: [all, verified, unverified, blocked] }
        - in: query
          name: country
          schema: { type: string }
        - in: query
          name: from
          schema: { type: string, format: date }
        - in: query
          name: to
          schema: { type: string, format: date }
      responses:
        '200':
          content:
            application/json:
              schema:
                allOf:
                  - { $ref: '#/components/schemas/Pagination' }
                  - type: object
                    properties:
                      data: { type: array, items: { $ref: '#/components/schemas/User' } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
...
  /api/admin/users/{userId}/block:
    patch:
      summary: Block a user account
      tags: [Admin]
      parameters:
        - in: path
          name: userId
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reason]
              properties:
                reason: { type: string }
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
  /api/admin/users/{userId}/unblock:
    patch:
      summary: Unblock a user account
      tags: [Admin]
      parameters:
        - in: path
          name: userId
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
...
  /api/admin/sessions:
    get:
      summary: List exam sessions
      tags: [Admin]
      parameters:
        - in: query
          name: page
          schema: { type: integer, default: 1 }
        - in: query
          name: limit
          schema: { type: integer, default: 25 }
        - in: query
          name: status
          schema: { type: string, enum: [generating, ready, active, completed, expired, aborted] }
        - in: query
          name: track
          schema: { type: string, enum: [associate, professional] }
        - in: query
          name: suspicious
          schema: { type: boolean }
        - in: query
          name: search
          schema: { type: string, description: "User email or session UUID" }
      responses:
        '200':
          content:
            application/json:
              schema:
                allOf:
                  - { $ref: '#/components/schemas/Pagination' }
                  - type: object
                    properties:
                      data: { type: array, items: { $ref: '#/components/schemas/ExamSession' } }
...
  /api/admin/certificates:
    get:
      summary: List certificates
      tags: [Admin]
      parameters:
        - in: query
          name: page
          schema: { type: integer, default: 1 }
        - in: query
          name: search
          schema: { type: string }
        - in: query
          name: status
          schema: { type: string, enum: [active, revoked] }
      responses:
        '200':
          content:
            application/json:
              schema:
                allOf:
                  - { $ref: '#/components/schemas/Pagination' }
                  - type: object
                    properties:
                      data: { type: array, items: { $ref: '#/components/schemas/Certificate' } }
  /api/admin/certificates/{certId}/revoke:
    patch:
      summary: Revoke a certificate
      tags: [Admin]
      parameters:
        - in: path
          name: certId
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reason]
              properties:
                reason: { type: string }
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
...
  /api/admin/flags:
    get:
      summary: List question flags (grouped by question_id)
      tags: [Admin]
      parameters:
        - in: query
          name: status
          schema: { type: string, default: open, enum: [open, reviewed, dismissed, escalated] }
        - in: query
          name: track
          schema: { type: string, enum: [associate, professional] }
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        question_id:   { type: string }
                        question_stem: { type: string }
                        flag_count:    { type: integer }
                        track:         { type: string }
                        thread:        { type: integer }
                        status:        { type: string }
                        last_flagged:  { type: string, format: date-time }
...